Refactor RecipeDrawer from RecipeDisplay
